composio-maxia/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "composio-maxia"
version = "0.1.0"
description = "MAXIA AI marketplace tools for Composio — swap crypto, rent GPUs, DeFi yields, AI services on 15 blockchains"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "MAXIA", email = "ceo@maxiaworld.app"}]
keywords = ["composio", "maxia", "ai-marketplace", "crypto", "defi", "gpu", "blockchain", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "composio-core>=0.6.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://maxiaworld.app"
Repository = "https://github.com/majorelalexis-stack/maxia"

[tool.setuptools.packages.find]
where = ["src"]
